expanding testing on retrieveObjectId
