+++ Sources/AutomergeSwiftAdditions/Document+retrieveObjectId.swift
