+++ Tests/AutomergeSwiftAdditionsTests/Document+retrieveObjectIdTests.swift
@@ -16,6 +16,12 @@ final class RetrieveObjectIdTests: XCTestCase {
     override func setUp() {
         setupCache = [:]
         doc = Document()
+
+        let _ = try! doc.put(obj: ObjId.ROOT, key: "name", value: .String("joe"))
+
+        let description = try! doc.putObject(obj: ObjId.ROOT, key: "description", ty: .Text)
+        setupCache["description"] = description
+
         let list = try! doc.putObject(obj: ObjId.ROOT, key: "list", ty: .List)
         setupCache["list"] = list
 
@@ -27,16 +33,88 @@ final class RetrieveObjectIdTests: XCTestCase {
         setupCache["deeplyNestedText"] = deeplyNestedText
     }
 
+    func testSetupDocPath() throws {
+        let pathToText = try! doc.path(obj: setupCache["deeplyNestedText"]!).stringPath()
+        XCTAssertEqual(setupCache.count, 4)
+        XCTAssertEqual(pathToText, ".list.[0].notes")
+    }
+
     func testPathAtRoot() throws {
         let doc = Document()
         let path = try! doc.path(obj: ObjId.ROOT)
         XCTAssertEqual(path, [])
     }
 
-    func testSetupDocPath() throws {
-        let pathToText = try! doc.path(obj: setupCache["deeplyNestedText"]!).stringPath()
-        XCTAssertEqual(setupCache.count, 3)
-        XCTAssertEqual(pathToText, ".list.[0].notes")
+    func testOverrideStrategy() throws {
+        let result = doc.retrieveObjectId(path: [AnyCodingKey("list")], containerType: .Key, strategy: .override)
+        switch result {
+        case let .success(success):
+            XCTFail("Expected override not implemented error, received: \(success)")
+        case .failure:
+            break
+        }
+    }
+
+    func testInvalidIndexLookup() throws {
+        let result = doc.retrieveObjectId(path: [], containerType: .Index, strategy: .readonly)
+        switch result {
+        case let .success(success):
+            XCTFail("Expected invalid lookup error, received: \(success)")
+        case .failure:
+            break
+        }
+    }
+
+    func testInvalidValueLookup() throws {
+        let result = doc.retrieveObjectId(path: [], containerType: .Value, strategy: .override)
+        switch result {
+        case let .success(success):
+            XCTFail("Expected invalid lookup error, received: \(success)")
+        case .failure:
+            break
+        }
+    }
+
+    func testLookupThroughText() throws {
+        let result = doc.retrieveObjectId(
+            path: [AnyCodingKey("description"), AnyCodingKey("list")],
+            containerType: .Value,
+            strategy: .readonly
+        )
+        switch result {
+        case let .success(success):
+            XCTFail("Expected invalid lookup error, received: \(success)")
+        case .failure:
+            break
+        }
+    }
+
+    func testLookupThroughScalar() throws {
+        let result = doc.retrieveObjectId(
+            path: [AnyCodingKey("joe"), AnyCodingKey("age")],
+            containerType: .Value,
+            strategy: .readonly
+        )
+        switch result {
+        case let .success(success):
+            XCTFail("Expected invalid lookup error, received: \(success)")
+        case .failure:
+            break
+        }
+    }
+
+    func testLookupBeyondBounds() throws {
+        let result = doc.retrieveObjectId(
+            path: [AnyCodingKey("list"), AnyCodingKey(2), AnyCodingKey("name")],
+            containerType: .Value,
+            strategy: .readonly
+        )
+        switch result {
+        case let .success(success):
+            XCTFail("Expected invalid lookup error, received: \(success)")
+        case .failure:
+            break
+        }
     }
 
     func testRetrieveLeafValue() throws {
